refactor: ops
