+++ src/utils/operator.ts
